src/requirements.txt
requests
tinydb
sqlalchemy 
psycopg2-binary
python-dotenv